requirements.txt
loguru==0.7.0
matplotlib==3.1.1
open_clip_torch==2.19.0
Pillow
streamlit==1.22.0
torch==1.13.1
tqdm==4.28.1